+++ apps/vscode/src/providers/webview.ts
@@ -49,9 +49,8 @@ export class QuartoWebviewManager<T extends QuartoWebview<S>, S> {
 
     context.subscriptions.push(
       window.registerWebviewPanelSerializer(this.viewType_, {
-        deserializeWebviewPanel: async (panel, _state) => {
-          // cleanup zombie preview
-          panel.dispose();
+        deserializeWebviewPanel: async (panel, state: S) => {
+          this.restoreWebvew(panel, state);
         },
       })
     );
@@ -100,14 +99,12 @@ export class QuartoWebviewManager<T extends QuartoWebview<S>, S> {
     }
   }
 
-  /*
-  private restoreWebvew(panel: WebviewPanel, state: any): void {
-    const url = state?.url ?? "";
-    const view = new this.webviewType_(this.extensionUri_, url, panel);
+  private restoreWebvew(panel: WebviewPanel, state: S): void {
+    const view = new this.webviewType_(this.context, state, panel);
     this.registerWebviewListeners(view);
     this.activeView_ = view;
   }
-  */
+
 
   private createWebview(
     context: ExtensionContext,